Working can

+++ DY-6MW7U59-3_WT32i_Bluetooth/can.h
@@ -16,7 +16,7 @@ enum beeps
 	LONG_BEEP_3 = 0x10
 };
 
-float GetLitersPerHour();
+uint16_t GetLitersPerHour();
 
 void CanRxHandler(CanRxMsg* RxMessage);
